inital bench update
