+++ content/english/blog/spiking-neural-network-framework-benchmarking/data.csv
@@ -1,61 +1,19 @@
-,framework,neurons,pass,time [s]
-0,Spyx (fp16)<br> v0.1.9,512,forward,0.0008169346899341788
-1,Spyx (fp16)<br> v0.1.9,4096,forward,0.010695806209077226
-2,Spyx (fp16)<br> v0.1.9,8192,forward,0.02386324826408835
-3,Spyx (fp32)<br> v0.1.9,512,forward,0.001162012002747787
-4,Spyx (fp32)<br> v0.1.9,4096,forward,0.018527232178854287
-5,Spyx (fp32)<br> v0.1.9,8192,forward,0.034034236272176104
-6,SpikingJelly CuPy<br>v0.0.0.0.15,512,forward,0.00208638109678521
-7,SpikingJelly CuPy<br>v0.0.0.0.15,4096,forward,0.02017000546822181
-8,SpikingJelly CuPy<br>v0.0.0.0.15,8192,forward,0.02700501052956832
-9,Rockpool<br>v2.7,512,forward,0.450988245010376
-10,Rockpool<br>v2.7,4096,forward,0.5514374375343323
-11,Rockpool<br>v2.7,8192,forward,0.8179528713226318
-12,Rockpool EXODUS<br>v2.7,512,forward,0.0063792095062839
-13,Rockpool EXODUS<br>v2.7,4096,forward,0.17546647787094116
-14,Rockpool EXODUS<br>v2.7,8192,forward,0.5170656442642212
-15,Sinabs<br>v1.2.9,512,forward,0.08140590667724609
-16,Sinabs<br>v1.2.9,4096,forward,0.15627563916719878
-17,Sinabs<br>v1.2.9,8192,forward,0.45073709487915037
-18,Sinabs EXODUS<br>v1.1.2,512,forward,0.005014166796117797
-19,Sinabs EXODUS<br>v1.1.2,4096,forward,0.1618201365837684
-20,Sinabs EXODUS<br>v1.1.2,8192,forward,0.5298179984092712
-21,snnTorch<br>v0.7.0,512,forward,0.15709185600280762
-22,snnTorch<br>v0.7.0,4096,forward,0.2415475845336914
-23,snnTorch<br>v0.7.0,8192,forward,0.6590602397918701
-24,Norse<br>v1.0.0,512,forward,0.17225990692774454
-25,Norse<br>v1.0.0,4096,forward,0.22791510158114964
-26,Norse<br>v1.0.0,8192,forward,0.5284585356712341
-27,Spyx (fp16)<br> v0.1.9,512,backward,0.0023113923668313776
-28,Spyx (fp16)<br> v0.1.9,4096,backward,0.016376035348743413
-29,Spyx (fp16)<br> v0.1.9,8192,backward,0.03882672663997201
-30,Spyx (fp32)<br> v0.1.9,512,backward,0.00331751147039019
-31,Spyx (fp32)<br> v0.1.9,4096,backward,0.02607223870604933
-32,Spyx (fp32)<br> v0.1.9,8192,backward,0.061416154196768095
-33,SpikingJelly CuPy<br>v0.0.0.0.15,512,backward,0.0021877023040271196
-34,SpikingJelly CuPy<br>v0.0.0.0.15,4096,backward,0.10078186988830566
-35,SpikingJelly CuPy<br>v0.0.0.0.15,8192,backward,0.025554166899787054
-36,Rockpool<br>v2.7,512,backward,0.4077626705169678
-37,Rockpool<br>v2.7,4096,backward,2.7636924584706626
-38,Rockpool<br>v2.7,8192,backward,5.59786057472229
-39,Rockpool EXODUS<br>v2.7,512,backward,0.008734763449456494
-40,Rockpool EXODUS<br>v2.7,4096,backward,0.07632951900876801
-41,Rockpool EXODUS<br>v2.7,8192,backward,0.09672652120175569
-42,Sinabs<br>v1.2.9,512,backward,0.3001319680895124
-43,Sinabs<br>v1.2.9,4096,backward,2.1311139265696206
-44,Sinabs<br>v1.2.9,8192,backward,4.366256554921468
-45,Sinabs EXODUS<br>v1.1.2,512,backward,0.012992295911235194
-46,Sinabs EXODUS<br>v1.1.2,4096,backward,0.06799119313557943
-47,Sinabs EXODUS<br>v1.1.2,8192,backward,0.09323820114135742
-48,snnTorch<br>v0.7.0,512,backward,0.10173825025558472
-49,snnTorch<br>v0.7.0,4096,backward,0.2020041227340698
-50,snnTorch<br>v0.7.0,8192,backward,0.723294734954834
-51,Norse<br>v1.0.0,512,backward,0.17903570334116617
-52,Norse<br>v1.0.0,4096,backward,1.2727914651234944
-53,Norse<br>v1.0.0,8192,backward,2.673928419748942
-54,Lava DL<br>v0.4.0.dev0,512,forward,0.0037799142441659606
-55,Lava DL<br>v0.4.0.dev0,4096,forward,0.08041971206665038
-56,Lava DL<br>v0.4.0.dev0,8192,forward,0.21762444972991943
-57,Lava DL<br>v0.4.0.dev0,512,backward,0.005953287084897359
-58,Lava DL<br>v0.4.0.dev0,4096,backward,0.07756125926971436
-59,Lava DL<br>v0.4.0.dev0,8192,backward,0.24242258071899414
+framework,neurons,forward,backward,memory
+Rockpool<br>v2.7,512,0.5370575189590454,0.3929328918457031,224404480
+Rockpool<br>v2.7,4096,0.6190436482429504,1.693268060684204,1843463168
+Rockpool<br>v2.7,8192,0.7726977666219076,1.8559740384419758,3810657280
+Rockpool EXODUS<br>v2.7,512,nan,nan,0
+Rockpool EXODUS<br>v2.7,4096,nan,nan,0
+Rockpool EXODUS<br>v2.7,8192,nan,nan,0
+snnTorch<br>v0.7.0,512,0.21930665969848634,0.12116854331072639,101212160
+snnTorch<br>v0.7.0,4096,0.2517068650987413,0.29547170230320524,858562560
+snnTorch<br>v0.7.0,8192,0.5316418409347534,0.7728563149770101,2131362816
+Lava DL<br>v0.4.0.dev0,512,0.002676320904716451,0.004092555348370948,151331840
+Lava DL<br>v0.4.0.dev0,4096,0.0413358211517334,0.022942618890242142,1255678976
+Lava DL<br>v0.4.0.dev0,8192,0.11928739267237046,0.09231607480482622,2909806080
+Spyx full-precision v0.1.9,512,0.017824094907372398,0.04740338287386185,nan
+Spyx full-precision v0.1.9,4096,0.4240721225738525,0.891455078125,nan
+Spyx full-precision v0.1.9,8192,1.6346806685129802,3.217236439387003,nan
+Spyx half-precision v0.1.9,512,0.03899453236506535,0.07789674579587758,nan
+Spyx half-precision v0.1.9,4096,0.6155067682266235,0.8670450448989868,nan
+Spyx half-precision v0.1.9,8192,1.9288682142893474,3.467363913853963,nan
